# ENH: Optimize performance of np.atleast_1d