--- a/benchmarks/benchmarks/bench_shape_base.py
+++ b/benchmarks/benchmarks/bench_shape_base.py
@@ -152,3 +152,19 @@ class Kron(Benchmark):
 
     def time_mat_kron(self):
         np.kron(self.large_mat, self.large_mat)
+
+class AtLeast1D(Benchmark):
+    """Benchmarks for np.atleast_1d"""
+
+    def setup(self):
+        self.x = np.array([1, 2, 3])
+        self.zero_d = np.float64(1.)
+
+    def time_atleast_1d(self):
+        np.atleast_1d(self.x, self.x, self.x)
+
+    def time_atleast_1d_reshape(self):
+        np.atleast_1d(self.zero_d, self.zero_d, self.zero_d)
+
+    def time_atleast_1d_single_argument(self):
+        np.atleast_1d(self.x)