In this PR we improve the performance of `np.atleast_1d` for the single argument case (and slightly improve for the other cases). 
The `np.atleast_1d` is used in many other methods as part of the input pre-processing ([Github search for np.atleast_1d](https://github.com/search?q=np.atleast_1d+language%3APython&type=code&l=Python)).

Benchmark
```
atleast_1d(1): Mean +- std dev: [main] 489 ns +- 16 ns -> [pr] 422 ns +- 9 ns: 1.16x faster
atleast_1d(x): Mean +- std dev: [main] 224 ns +- 9 ns -> [pr] 159 ns +- 7 ns: 1.41x faster
atleast_1d(1, x ): Mean +- std dev: [main] 592 ns +- 26 ns -> [pr] 605 ns +- 24 ns: 1.02x slower
polyadd: Mean +- std dev: [main] 2.50 us +- 0.11 us -> [pr] 2.26 us +- 0.08 us: 1.10x faster

Geometric mean: 1.15x faster
```

<details><summary>Benchmark script</summary>

```
import pyperf

runner = pyperf.Runner()

setup="""
import numpy as np
from numpy.polynomial import Polynomial

x = np.array([2, 3, 4])
y = np.array([2, 3])

p = Polynomial([1, 2, 3.])
"""   
runner.timeit(name=f"atleast_1d(1)", stmt=f"np.atleast_1d(1)", setup=setup)
runner.timeit(name=f"atleast_1d(x)", stmt=f"np.atleast_1d(x)", setup=setup)
runner.timeit(name=f"atleast_1d(1, x )", stmt=f"np.atleast_1d(1, x )", setup=setup)
runner.timeit(name=f"polyadd", stmt=f"np.polyadd(x, y)", setup=setup)
```
</details>
<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->