@ngoldbaum 2024-05-01
We looked at this at the triage meeting and think this needs asv benchmarks. It would be good to see that the slowdown for the list of arrays case is not significant. It would also be good to add a comment explaining that the single-element case is done first for performance reasons.