@eendebakpt 2024-05-02
> We looked at this at the triage meeting and think this needs asv benchmarks. It would be good to see that the slowdown for the list of arrays case is not significant. It would also be good to add a comment explaining that the single-element case is done first for performance reasons.

I added 3 benchmark functions. On my system there is no slowdown visible for the list of arrays case, but the asv runs are a bit noisy so the variation between main and this PR is typically smaller than the variations between different runs. @ngoldbaum Could you run the benchmarks independently?