@ngoldbaum 2024-05-06
Here's what I see running the new benchmarks on my laptop:

```
| Change   | Before [b2960879] <main>   | After [f6cb2a92] <atleast_1d>   |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------------|---------------------------------|---------|------------------------------------------------------------|
| -        | 206±10ns                   | 167±5ns                         |    0.81 | bench_shape_base.AtLeast1D.time_atleast_1d_single_argument |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```

So on my machine, about 20% faster and any decrease in performance on the list-of-arrays case is less than 5%. Seems worth it to me!